extractor/requirements.txt
# AI Credit Decisioning Engine — Extractor
# requirements.txt

# ── Core PDF extraction ─────────────────────────
pdfplumber>=0.10.0          # Text-based PDF extraction + table parsing

# ── Scanned PDF / OCR ──────────────────────────
pdf2image>=1.16.3           # Converts PDF pages to images for OCR
pytesseract>=0.3.10         # OCR engine wrapper
Pillow>=10.0.0              # Image processing for OCR pre-processing

# ── Document formats ────────────────────────────
python-docx>=1.1.0          # DOCX extraction
openpyxl>=3.1.2             # XLSX extraction

# ── LLM APIs ───────────────────────────────────
anthropic>=0.34.0           # Claude API (preferred)
openai>=1.40.0              # OpenAI GPT-4 (fallback)

# ── Utilities ───────────────────────────────────
python-dotenv>=1.0.0        # Load .env file for API keys